requirements.txt
streamlit
streamlit-autorefresh
hyperliquid-python-sdk
pandas
plotly
pyperclip